+++ app/viewmodels/CheckYourAnswersHelper.scala
@@ -182,58 +182,6 @@ class CheckYourAnswersHelper(userAnswers: UserAnswers)(implicit messages: Messag
       case _ => None
     }
   }
-
-  private def checkPage[A](page: QuestionPage[A])(implicit rds: Reads[A]): Option[Page] =
-    userAnswers.get(page) match {
-      case None => Some(page)
-      case _    => None
-    }
-
-  private def checkPrimaryContactDetails: Seq[Page] = Seq(
-    checkPage(ContactNamePage),
-    checkPage(ContactEmailPage)
-  ).flatten ++ checkPrimaryContactNumber
-
-  private def checkPrimaryContactNumber: Seq[Page] = (userAnswers.get(HaveTelephonePage) match {
-    case Some(true)  => checkPage(ContactPhonePage)
-    case Some(false) => None
-    case _           => Some(HaveTelephonePage)
-  }).toSeq
-
-  private def checkSecondaryContactPhone: Seq[Page] = (userAnswers.get(SecondContactHavePhonePage) match {
-    case Some(true)  => checkPage(SecondContactPhonePage)
-    case Some(false) => None
-    case _           => Some(SecondContactHavePhonePage)
-  }).toSeq
-
-  private def checkSecondaryContactDetails: Seq[Page] =
-    userAnswers.get(HaveSecondContactPage) match {
-      case Some(true) =>
-        Seq(
-          checkPage(SecondContactNamePage),
-          checkPage(SecondContactEmailPage)
-        ).flatten ++ checkSecondaryContactPhone
-      case Some(false) => Seq.empty
-      case _           => Seq(HaveSecondContactPage)
-    }
-
-  private def validate: Seq[Page] = checkPrimaryContactDetails ++ checkSecondaryContactDetails
-
-  private def pageToRedirectUrl(mode: Mode): Map[Page, String] = Map(
-    ContactNamePage            -> controllers.routes.ContactNameController.onPageLoad(mode).url,
-    ContactEmailPage           -> controllers.routes.ContactEmailController.onPageLoad(mode).url,
-    HaveTelephonePage          -> controllers.routes.HaveTelephoneController.onPageLoad(mode).url,
-    ContactPhonePage           -> controllers.routes.ContactPhoneController.onPageLoad(mode).url,
-    HaveSecondContactPage      -> controllers.routes.HaveSecondContactController.onPageLoad(mode).url,
-    SecondContactNamePage      -> controllers.routes.SecondContactNameController.onPageLoad(mode).url,
-    SecondContactEmailPage     -> controllers.routes.SecondContactEmailController.onPageLoad(mode).url,
-    SecondContactHavePhonePage -> controllers.routes.SecondContactHavePhoneController.onPageLoad(mode).url,
-    SecondContactPhonePage     -> controllers.routes.SecondContactPhoneController.onPageLoad(mode).url
-  )
-
-  def changeAnswersRedirectUrl(mode: Mode): Option[String] =
-    validate.headOption
-      .flatMap(pageToRedirectUrl(mode).get)
 }
 
 object CheckYourAnswersHelper {
